Added null and other checks to heal all and mute all

+++ Utils.cs
@@ -217,7 +217,7 @@ namespace MoreAdminCommands
         #region getDistance
         public static bool getDistance(Vector2 player, Vector2 mob, int radius)
         {
-            if (Math.Sqrt(((int)(mob.X - player.X) ^ 2) + ((int)(mob.Y - player.Y) ^ 2)) <= radius)
+            if ((int)Math.Sqrt(((int)(mob.X - player.X) ^ 2) + ((int)(mob.Y - player.Y) ^ 2)) <= radius)
                 return true;
             return false;
         }
